+++ snips/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 3.3
 - Updated to latest snips 0.58.3
+- NOTE: This version requires an updated assistant from the Snips Console.
 
 ## 3.2
 - Added snips-asr-injections service to allow dynamic entity names